refactor: create new files to improve readability
